Save and Reset Gain+Offset on Image capture

* save and reset gain on capture

* add gain to image metadata

* display gain from camera equipment by default instead of empty value

* fix nikon saving gain, add preselected and default gain instead of empty value, change gain to int where necessary

* Merge remote-tracking branch 'upstream/develop' into feature/gainreset

* Merge branch 'main' into feature/gainreset

* change short gain to int gain

* when camera is not connected and gain is displayed display (Camera) instead of DependencyProperty.UnsetValue


Former-commit-id: a6fc1f937bd913c342abda4c90f364fdd9e0f58b
Former-commit-id: c06a1067cbe85149a1f484cee160391ae39d7d9f

+++ NINA/Utility/Mediator/Interfaces/ICameraMediator.cs
@@ -37,7 +37,7 @@ namespace NINA.Utility.Mediator.Interfaces {
 
         void SetBinning(short x, short y);
 
-        void SetGain(short gain);
+        void SetGain(int gain);
 
         void SetSubSample(bool subSample);
